remove ast.h/.c
