+++ src/analyze/builtin.c
@@ -1,5 +1,4 @@
 #include "structs.h"
-#include "ast.h"
 #include "object.h"
 #include "main.h"
