+++ src/analyze/object.h
@@ -2,7 +2,6 @@
 #define _OBJECT_H
 
 #include "structs.h"
-#include "ast.h"
 
 typedef struct {
   ObjectHeader;
@@ -63,6 +62,44 @@ typedef struct {
 } BlockObj;
 
 BlockObj *newBlock(List *stats, struct ExpressionObj *e);
+
+enum ExpType {
+  ExpCmp,
+  ExpLogic,
+  ExpCalc,
+  ExpStr,
+  ExpNum,
+  ExpBool,
+  ExpVar,
+  ExpArray,
+  ExpFunc,
+};
+
+typedef enum ExpType ExpType;
+
+enum ExpVType {
+  TNone,
+  TPackage,
+  TSkill,
+  TGeneral,
+  TNumber,
+  TBool,
+  TString,
+  TPlayer,
+  TCard,
+  TEmptyList,
+  TPlayerList,
+  TCardList,
+  TNumberList,
+  TStringList,
+  TMark,
+  TFunc,
+
+  TNotSure = 0xFFFE,
+  TAny = 0xFFFF
+};
+
+typedef enum ExpVType ExpVType;
 
 typedef struct ExpressionObj {
   ObjectHeader;
