+++ src/parse/error.c
@@ -176,6 +176,37 @@ const char *yytr(const char *orig) {
   return orig;
 }
 
+static const char *type_table[] = {
+  "无类型",
+  "拓展包类型",
+  "技能类型",
+  "武将类型",
+  "数字类型",
+  "布尔类型",
+  "字符串类型",
+  "玩家类型",
+  "卡牌类型",
+  "空数组",
+  "玩家数组",
+  "卡牌数组",
+  "数字数组",
+  "字符串数组",
+  "标记类型",
+  "函数类型"
+};
+
+void checktype(void *o, ExpVType a, ExpVType t) {
+  if (a != t && a != TAny && t != TAny) {
+    if (!o) {
+      fprintf(error_output, "Type error: expect %d, got %d\n", t, a);
+    } else {
+      YYLTYPE *obj = o;
+      yyerror(obj, "类型不匹配：需要 %s，但得到的是 %s",
+              type_table[t], type_table[a]);
+    }
+  }
+}
+
 /*
  * fkparse处理的文本基本都是汉字
  * 一个汉字在UTF-8中是三个字节，但是输出到屏幕上面时候只有两个字母的宽度
