+++ src/parse/error.h
@@ -6,7 +6,7 @@
 #include "main.h"
 
 const char *yytr(const char *orig);
-
+void checktype(void *o, ExpVType a, ExpVType t);
 void yyerror(YYLTYPE *loc, const char *msg, ...);
 
 #endif
