+++ src/parse/grammar.y
@@ -1,7 +1,6 @@
 %{
 #include "main.h"
 #include "enums.h"
-#include "ast.h"
 #include "object.h"
 #include "grammar.h"
 #include "error.h"
